Several improvements in create_map.sh

+++ garmin/scripts/create_map.sh
@@ -1,20 +1,19 @@
 #!/bin/bash
 
-# Create a Garmin map for the specified region.
-#
-# Usage:
-#
-# ./create_map.sh REGIONURL MAPID
-#
-# REGIONURL must refer to a *.osm.pbf file on https://download.geofabrik.de/
-# MAPID must be a unique 4-digit number, e.g., based on the telephone country code
-#
-# Examples:
-#
-# ./create_map.sh https://download.geofabrik.de/europe/alps-latest.osm.pbf 4100
-# ./create_map.sh https://download.geofabrik.de/europe/germany-latest.osm.pbf 4900
-# ./create_map.sh https://download.geofabrik.de/europe/united-kingdom.html 4400
-#
+display_usage() {
+    echo -e "\nCreate a map for Garmin GPS devices using OpenTopoMap styles.\n"
+    echo -e "Usage:"
+    echo -e "$0 REGIONURL MAPID\n"
+    echo -e "REGIONURL must refer to a *.osm.pbf file on https://download.geofabrik.de/"
+    echo -e "MAPID must be a unique 4-digit number, e.g., based on the telephone country code\n"
+    echo -e "Example:"
+    echo -e "$0 https://download.geofabrik.de/europe/germany-latest.osm.pbf 4900"
+}
+
+if [ $# -le 1 ]; then
+  display_usage
+  exit 1
+fi
 
 REGIONURL=$1
 MAPID=$2
@@ -36,43 +35,30 @@ if ! [[ $MAPID =~ $re ]] ; then
 fi
 
 REGION=$(echo ${REGIONURL} | sed 's#.*/##g' | sed 's#\.osm\.pbf##g')
-
-echo "########## Creating map for region ${REGION} ##########"
-
-# Go to the OpenTopoMap/garmin folder
-cd $(dirname "$0")/..
-
-# Adjust the versions of these two tools to their latest relase,
-# according to www.mkgmap.org.uk
-MKGMAP="mkgmap-r4919"
-SPLITTER="splitter-r654"
+echo -e "\n########## Create map for region ${REGION} ##########"
 
 #
-# Download mkgmap, splitter, bounds, and sea
+# Change to the folder where the script is located and set env variables.
 #
-mkdir -p tools
-pushd tools > /dev/null
+cd $(dirname $0) || { echo "cd command failed"; exit 1; }
+source setup_vars.bash
 
-if [ ! -d "${MKGMAP}" ]; then
-    echo "Downloading ${MKGMAP}..."
-    wget "http://www.mkgmap.org.uk/download/${MKGMAP}.zip"
-    unzip "${MKGMAP}.zip"
-else
-    echo "${MKGMAP} already downloaded"
-fi
-MKGMAPJAR="$(pwd)/${MKGMAP}/mkgmap.jar"
+#
+# Download required tools
+#
+echo -e "\n##### Download required tools #####"
+./download_tools.sh
 
-if [ ! -d "${SPLITTER}" ]; then
-    echo "Downloading ${SPLITTER}..."
-    wget "http://www.mkgmap.org.uk/download/${SPLITTER}.zip"
-    unzip "${SPLITTER}.zip"
-else
-    echo "${SPLITTER} already downloaded"
-fi
-SPLITTERJAR="$(pwd)/${SPLITTER}/splitter.jar"
+# Go to the OpenTopoMap/garmin folder.
+cd .. || { echo "cd command failed"; exit 1; }
 
-popd > /dev/null
+MKGMAPJAR="$(pwd)/tools/${MKGMAP}/mkgmap.jar"
+SPLITTERJAR="$(pwd)/tools/${SPLITTER}/splitter.jar"
 
+#
+# Download bounds
+#
+echo -e "\n##### Download bounds and sea #####"
 if stat --printf='' bounds/bounds_*.bnd 2> /dev/null; then
     echo "bounds already downloaded"
 else
@@ -81,9 +67,15 @@ else
     wget -O bounds.zip "http://osm.thkukuk.de/data/bounds-latest.zip"
     unzip "bounds.zip" -d bounds
 fi
-
 BOUNDS="$(pwd)/bounds"
+if [ ! -d "${BOUNDS}" ]; then
+    echo "Failed to extract ${BOUNDS} folder."
+    exit 1
+fi
 
+#
+# Download sea
+#
 if stat --printf='' sea/sea_*.pbf 2> /dev/null; then
     echo "sea already downloaded"
 else
@@ -92,13 +84,16 @@ else
     wget -O sea.zip "https://www.thkukuk.de/osm/data/sea-latest.zip"
     unzip "sea.zip"
 fi
-
 SEA="$(pwd)/sea"
+if [ ! -d "${SEA}" ]; then
+    echo "Failed to extract ${SEA} folder."
+    exit 1
+fi
 
 #
-# Fetch map data, split & build garmin map
+# Fetch map data
 #
-
+echo -e "\n##### Fetch map data #####"
 mkdir -p data
 pushd data > /dev/null
 
@@ -108,8 +103,15 @@ else
     echo "downloading ${REGION}.osm.pbf"
     wget ${REGIONURL}
 fi
+if [ ! -f ${REGION}.osm.pbf ]; then
+    echo "Failed to download ${REGION}.osm.pbf."
+    exit 1
+fi
 
-echo "##### Running ${SPLITTERJAR}... #####"
+#
+# Run splitter
+#
+echo -e "\n##### Running ${SPLITTERJAR}... #####"
 rm -f ${MAPID}*.pbf areas.* densities-out.txt template.args
 java -Xmx8192M -jar $SPLITTERJAR --precomp-sea=$SEA  --mapid=${MAPID}0001 "$(pwd)/${REGION}.osm.pbf"
 DATA="$(pwd)/${MAPID}*.pbf"
@@ -119,21 +121,25 @@ popd > /dev/null
 OPTIONS="$(pwd)/opentopomap_options"
 STYLEFILE="$(pwd)/style/opentopomap"
 
+#
+# Generate opentopomap.typ
+#
+echo -e "\n##### Generating opentopomap.typ... #####"
 pushd style/typ > /dev/null
-
-echo "##### Generating opentopomap.typ... #####"
 java -jar $MKGMAPJAR --family-id=35 opentopomap.txt
 TYPFILE="$(pwd)/opentopomap.typ"
-
 popd > /dev/null
 
-echo "##### Running ${MKGMAPJAR}... #####"
+#
+# Run mkgmap
+#
+echo -e "\n##### Running ${MKGMAPJAR}... #####"
 rm -rf ./output_tmp
 java -Xmx8192M -jar $MKGMAPJAR -c $OPTIONS --style-file=$STYLEFILE \
     --precomp-sea=$SEA \
     --dem=./dem1/all \
     --output-dir=output_tmp --bounds=$BOUNDS $DATA $TYPFILE
 
-echo "Generated map will be saved as output/otm-${REGION}.img"
+echo -e "\nDone! Generated map will be saved as output/otm-${REGION}.img"
 mkdir -p output
 mv output_tmp/gmapsupp.img output/otm-${REGION}.img
